Split large atoms file into smaller ones
